requirements.txt
Flask
bcrypt
gunicorn
flask-cors
PyJWT
python-dotenv
flask_socketio
gevent>=1.4
flask_sqlalchemy
flask_jwt_extended
werkzeug
requests
dnspython
